[rust png] Account for `fFrameAtCurrentStreamPosition` inaccuracies.
`int fFrameAtCurrentStreamPosition` gives the number of the frame for
which the `IDAT` chunk sequence starts at the current input stream
position, **or before** the current input stream position.  Before this
CL `SkPngRustCodec::startDecoding` and
`SkPngRustCodec::seekToStartOfFrame` wouldn't correctly account for the
"or before" part, which meant that they could start decoding from the
middle of a frame.

After this CL, tracking of the stream position is augmented with a new
`fStreamIsPositionedAtStartOfFrameDat` field.

To catch similar issues in the future, the `DecodingState::fDst` span is
tightened (to only span the current frame, not till the end of the whole
image) and a new assertion is added to ensure that when the Rust codec
reports "no more rows", then the whole `fDst` has indeed been filled
out. And this CL also adds a regression test.

Bug: chromium:356922876
Change-Id: I898c15b829900f162ec21e816a9a620f4de32a60
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/913879
Commit-Queue: Łukasz Anforowicz <lukasza@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>