SkCrabbyAvifCodec: Add support for gainmaps
Add onGetGainmapCodec callback to SkCrabbyAvifCodec. It makes a
copy of the AVIF data and asks the underlying decoder to return
only the gainmap images.

It also populates the SkGainmapInfo from the underlying
crabbyavif::avifImage.

fGainmapMathColorSpace is not yet computed and it will come in a
follow-up.

Bug: b/306268460
Change-Id: I0414eaf60236b4d8a9aae847cb477925b3c458a5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/914476
Commit-Queue: Vignesh Venkat <vigneshv@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>