[pdf] Augment glyph to unicode map
Currently the ToUnicode CMap is constructed by creating a glyph to
unicode mapping based on the character map of the typeface (with
conflicts being resolved arbitrarially) and emitting the mappings for
the glyphs that are actually used. Anything else is wrapped in a
marked-content sequence with an /ActualText entry in its property list.

Add to this an additional extended mapping from glyph to unicode
strings which is added to as such mappings are encountered. The common
use of this is to allow substitute glyphs or ligatures to be mapped
through ToUnicode.

This new mapping allows for additional mappings to be added, but does
not solve the issue where the initial typeface based mapping resolves
arbitrarially and this cannot be changed when it might be better to pick
a different mapping.

Bug: 40043392
Change-Id: I90ff87532770c5f564b2156cfd30d5c08384acb6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/914976
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>