[Graphite] Use PushConstants instead of VBO for load msaa draws.
For protected mode we can't have a protect VBO. However, for the MSAA
load draws we currently depend on using vkCmdUpdateBuffer to update the
VBO. This command would require the buffer to be protected which it
cannot be.

This change instead changes the shader to upload the bounds of the load
as uniforms. We don't need a VBO and instead infer the vertex positions
by the vertexIDs. To update the uniforms for this shader we use
pushConstants to avoid the protected issues.

Change-Id: Ia87d464465b17085b48a7fa84058483bea7d2cb7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/914436
Commit-Queue: Greg Daniel <egdaniel@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>