Extend the lifetime of IntrinsicConstantsManager
Currently IntrinsicConstantsManager resides in DawnCommandBuffer, which
causes the rebuilding of the intrinsic cache every time a command buffer
is sumitted and recreated. This incurs more cache misses and dawn buffer
writes for intrinsic constants. To mitigate this, this moves it to
DawnResourceProvider, so the cache can preserve across command buffer
submissions.

Bug: chromium:370252259
Change-Id: I34864f7cee2b64fe6389a907ab5cdc29260cb2ae
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/911437
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>