Guard ganesh dependency from debugger
Building canvaskit with debugger was broken because
of the testonly=true. This had been added in http://review.skia.org/895117
because DM depends on debugger but needs GL support with the
test defines. One solution would have been to make canvaskit
"testonly" but that feels wrong.

The proper solution (and one we might have to do if debugger
stops working) is refactor debugger to allow it to depend on the GPU.

Let's try this and see if it unbreaks debugger.skia.org.

Change-Id: I4ecfa2ef703e7c6bfffcb5490b12871c13fefd26
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/917677
Reviewed-by: Florin Malita <fmalita@google.com>
Auto-Submit: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>