[rust png] Fix memory safety issue by copying `iCCP` chunk data.
`skcms_Parse` forwards to `skcms_ParseWithA2BPriority` which explicitly
documents that "the buffer is not copied; it must remain valid as long
as the skcms_ICCProfile will be used".  Before this CL, this lifetime
requirement can be violated - the `iCCP` chunk data is owned by
`SkPngRustCodec::fReader` but the reader may be reset when
`SkPngRustCodec::seekToStartOfFrame` rewinds the stream.  This CL fixes
this by explicitly working with an owned copy of the data.

Bug: chromium:356884909
Change-Id: Ib281f682f153dc3f1336ec00d526c13200d33ede
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/917191
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Commit-Queue: Łukasz Anforowicz <lukasza@google.com>