[viewer] QOL improvements to perspective and zoom
The clipRect() was originally intended to make the real perspective
function like the fake perspective mode, but tends to limit exploration
of rendering under perspective and greatly interferes with slides that
reset the CTM for some UI around the rest of the perspective content.

The black outline around the zoom widget's focused pixel could be
very hard to identify when looking at dark content so this switches
it to be responsive to the zoomed content and have a high contrast
in all scenarios.

Change-Id: Iada9e5ecd219f65b9001259105141ea09cb29e20
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/920616
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>