Subset SkImage_Picture objects instead of rendering to Bitmap image.
When rendering a non-GPU backed tiled image we would normally
render the image to a bitmap and then copy regions out into tiled
images in an ImageSet. However, if we have a Picture-backed image
then we might have GPU-only data in there which can only be rendered
on a GPU Device. The solution is to create subsets of the Picture-
backed image to represent the tiles and insert those in the ImageSet.
The subsets in turn are created by creating a new Picture-backed
image rather than (again) rendering to a bitmap and subsetting that.

One side effect of this ended up scaling the local coordinates to a
a degree that they couldn't be represented in a half-precision float,
leading to banding in gradient calculations. To solve this I changed
the t calculations for gradients in Ganesh to use float instead of half.

Bug: b/372765411
Change-Id: I3f68a2ed5eefabd31eaa35a573ced10f682caba7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/919696
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>