[graphite] Add PrecompileContext object
This CL introduces an object that can be created on the main thread and then moved to a precompilation thread. This is substantially safer than passing the Context to a precompilation thread.

It also holds a ResourceProvider so eliminates the need to allocate a new one for each Precompile call.

Bug: b/358074434
Bug: b/373849852
Change-Id: I4ba4530655d65224dcd68e1b1a3e2a0d477bd5fd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/919738
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>