Change window::DisplayParams to be immutable and passed by pointer
The #ifdef in DisplayParams is causing issues when trying to build
viewer with both Ganesh and Graphite. To remove that, we need to
introduce polymorphism, which means changing all the uses of DisplayParams
to be a pointer. In order to prevent threading issues, we use std::unique_ptr<const DisplayParams> and then hide the fields
only allowing the fields read-only access.

To facilitate some uses where small tweaks are made to the params,
a DisplayParamsBuilder is added to easily clone and modify them.

Future CLs will add a subclass for the graphite version of params
and move that into graphite-only code which will remove the
problematic #ifdefs.

Suggested Review Order:
 - tools/window/DisplayParams.h
 - tools/sk_app/Window.* and then tools/sk_app/**/Window_*
 - tools/window/WindowContext.*
 - tools/window/VulkanWindowContext.* noting the removal of
   createSwapchain's third parameter which seemed unnecessary
   (it was always passing in fDisplayParams and then setting it
    to itself)
 - The rest of tools/window/**
 - all other files in any order

Change-Id: If593cdacacae52db9285445a236fd31a974f636e
Bug: b/256870320
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/908937
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>