Minor cleanups with AutoSTArray and AutoSTMalloc
AutoSTArray can be 4 bytes smaller if, for example,
sizeof(T) == sizeof(int) and kCount is odd
or sizeof(T) < sizeof(int) and *handwave*. This comes up
infrequently, but could be a way to trim a tiny bit of
space on allocations that come up often, or give us a
way to get one more "free" allocation on the stack.

There's also some consts that could be constexprs.

Change-Id: I6799395cf7c91d3018afb9aba485113042c8f763
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/924737
Reviewed-by: Michael Ludwig <michaelludwig@google.com>