Add a maxSize parameter to storeVkPipelineCacheData
In the case that the client has an upper limit on the amount of pipeline
data they can cache (which should be the case more often than not), the
driver can cleanly emit at most that much data.

In particular this API is required by HWUI.

For clients which continue using the unlimited overload, use SIZE_MAX as
the limit, since any data buffer larger than this would be not be fully
addressable anyway.

Note that, in the case that the driver had more data to output, the
return code from vkGetPipelineCacheData is VK_INCOMPLETE, so handle this
as a successful case.

Bug: b/269117286
Change-Id: I3ae889c12e6b1b70692d5f80b1940cba8225cc14
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/927636
Reviewed-by: Michael Ludwig <michaelludwig@google.com>