Set FragmentState::constants
FragmentState::constants is not initialized by default so it can point
to arbitrary memory. This probably doesn't matter if constantCount is 0,
since constants value shouldn't be read from, but the uninitialized
value is copied internally in dawn at [1]. That is triggering a MSAN
failure.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/dawn/src/dawn/native/RenderPipeline.cpp;l=944-946;drc=c58c3aa29319877e2e64d0bad3216f351ecb7e5a

Bug: chromium:381109466
Change-Id: Ie6a32dd6764a07266f0e4a0b7a82a8d10dc5f172
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/926982
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Kyle Charbonneau <kylechar@chromium.org>