Fix style and axis scanning for variable fonts
Have SkFontScanner::scanInstance be able to return the variation
position of the instance. This allows the user to then merge in any
requested variation position. Currently FreeType is getting the
incorrect style and variation position for variable instances. This was
being hidden by SkTypeface_FreeType::FaceRec::setupAxes ignoring any
provided variable position when creating a named instance. This change
also removes that.

This re-enables the FontConfigInterface_MatchStyleNamedInstance test
which was the only one requiring that the named instances and the style
be correct. This forces a fix to the Skia side Fontations code.

Change-Id: I8f375fc1d39f639b65c17bf6292776a391ff9901
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/926983
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>