Fix SkOrderedFontMgr::onMatchFamily
SkOrderedFontMgr's `matchFamily` currently has [a bug](https://issues.skia.org/issues/382016481) where it will only look in the first FontMgr in its list when searching for fonts. The problem is that its search loop is expecting a `null` response to signal "no matches", but the actual behavior of FontMgr is to never return null. Instead it will return a zero-length SkFontStyleSet.

This PR adds a `count() > 0` check to the return value rather than just checking whether it is non-null.

This is an imported pull request from
https://github.com/google/skia/pull/193

Added the author to the list of Skia AUTHORS

GitOrigin-RevId: 736a0e3705b5e05f680dd942dfca27f714ba1f91
Change-Id: I6e0f9a6fe5a5d7e6a45567bc0f83ed27024cb584
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/926716
Commit-Queue: Julia Lavrova <jlavrova@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: Julia Lavrova <jlavrova@google.com>