[graphite] Expand logging/tracing of significant precompile events
This is the majority of the logging/tracing changes. A follow on CL will add the pipeline-cache purging logging.

The new verbose pipeline logs are all structured as:

   <event-name><P or N>
   key <unique-key's-hash-value>
   compilationID <globally-unique-int>

Where the 'P' or 'N' encodes whether the Pipeline is from Precompilation or Normal compilation.

I'll document this somewhere else but, the <event-name>s form a finite state machine of the Pipeline's lifetime.

We have a compilationID (in addition to the pipeline key's hash) bc a given key can be compiled multiple times (due to the caching behavior, threading, and purging). Not accounting for that in the logging can be confusing.

The logging is controlled by the SK_PIPELINE_LIFETIME_LOGGING #define in SkUserConfig.h. The plan is to have this defined in Chrome's SkUserConfig.h but we also have it separately added to our local Debug Graphite builds for compilation coverage.

Bug: b/358074434
Change-Id: I2e5e8f96e76b2710f8b9fd973001d7ac4b1c67e6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/927979
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>