Replace SkSharedMutex with std::shared_mutex
We tried to do this before [1] but had to roll it back [2] because
apparently this didn't build for Flutter. I see references to
shared_mutex in Flutter's codebase, so hopefully we are safe to reland
this.

I noticed this while evaluating a patch to SkSharedMutex for gcc;
if we can get rid of SkSharedMutex, then we shouldn't need the patch.

I had to get rid of the thread annotations [3] because those don't seem
to work with stuff in the stl.

[1] http://review.skia.org/285691
[2] http://review.skia.org/286124
[3] https://clang.llvm.org/docs/ThreadSafetyAnalysis.html

Change-Id: I0d0e501ed60f8855fe6650237093b9ecb67840dc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/930127
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>