Allow getPixels to succeed for progressive JPEG's with no EOI marker
Previously, progressive Jpeg's that do not contain the EOI marker were not able to be decoded by Skia with the getPixels() function. This is common for progressive images to be missing, but are displayed through chrome and other browsers because they have enough data to be decoded progressively.
This change checks for progressive images and if they can be scanned at least once (producing an image the size of dst), the function will succeed. This matches previous behavior with non-progressive images in which they should not succeed if their data is truncated to a point smaller than the dst.
For progressive Jpeg's with EOI marker's, this doesn't change behavior as the default libjpeg functions we used before just combined all the scans together as this does.

For reference: https://github.com/libjpeg-turbo/libjpeg-turbo/blob/20ade4dea9589515a69793e447a6c6220b464535/doc/libjpeg.txt#L2029


Bug: 40043594
Change-Id: Id08225689e290b93fbca9757064c8eafe25e9fef
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/928441
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>