[rust png] Update `cxx` from 1.0.131 to 1.0.135.
This CL updates the `WORKSPACE.bazel` dependency on `cxx` and
`cxxbridge-cmd` crates to version `1.0.135`.

This CL is motivated by the desire to stop suppressing warnings in the
proc-macro-generated code, because in the new vesion that code should
be clean of warnings (after https://github.com/dtolnay/cxx/pull/1415).

The following commands have been used to mirror the crates:

$ go run ./bazel/gcs_mirror/gcs_mirror.go \
    --url https://crates.io/api/v1/crates/cxxbridge-cmd/1.0.135/download \
    --sha256 4717c9c806a9e07fdcb34c84965a414ea40fafe57667187052cf1eb7f5e8a8a9 \
    --add_suffix=.tar.gz
$ go run ./bazel/gcs_mirror/gcs_mirror.go \
    --url https://crates.io/api/v1/crates/cxx/1.0.135/download \
    --sha256 4d44ff199ff93242c3afe480ab588d544dd08d72e92885e152ffebc670f076ad \
    --add_suffix=.tar.gz

Bug: chromium:383521545
Change-Id: I713067e42c87384a76f42773adaf36a834a25735
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/930418
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Łukasz Anforowicz <lukasza@google.com>