[graphite] Adjust raster path key to include clip origin.
The current raster path key includes the mask size and the
fractional translation, but masks can represent a clip of the shape,
with an integer translation that's not reflected in the key. This CL
adds the origin of the mask relative to the initial transformed
shape bounds to distinguish between these cases.

Bug: b/384413092
Change-Id: I3269651428f6ee8565fab217df6ef4435684100e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/931827
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>