Remove CoreText 14 and 15 work arounds
kCTFontSymbolicTrait was already not used in CoreText < 14 (macOS 10.13
and earlier) and so the code using it can now be removed.

The CoreText 15 check was added as a work around in
https://skia-review.googlesource.com/c/skia/+/260346 for
https://issues.chromium.org/issues/40104990 which was later actually
fixed with https://issues.chromium.org/issues/40115836 so this
particular check isn't really necessary anymore, especially with the
change to using 0.07 instead of 1. Chromium no longer supports macOS
10.15, though some of Skia's other dependents claim to support it.
Removing this check doesn't remove support but might lead to
"interesting behavior" in somewhat uncommon situations only on macOS
10.15.

Change-Id: I25dfe22dc3c823fa61537c38316d526c194c2f71
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/931946
Reviewed-by: Florin Malita <fmalita@google.com>