Remove unneeded resource refs on Vulkan CommandBuffer.
All of the textures and buffers that we currently track on the
command buffer are already added to the command buffer from other parts
of the code. So there is no need to additionally add them from
the VulkanCommandBuffer.

Additionally this fixes issues where the VulkanCB was adding full
usage refs when it should have just ben CB refs.

Bug: b/361875857
Change-Id: I1d7469870289342300def9d3390df547721dcf92
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/931948
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>