Keep last resort glyphs and positions together
In the SkGlyphRunListPainterCPU::drawForBitmapDevice last resort case it
is possible that even after attempting to get the glyphs to fit it may
still not be possible to create a mask for a glyph and the glyph will be
rejected. Currently if this happens the `sourcePositions` and `accepted`
will no longer be syncronized, possibly leading to out of bounds reads.

Because the bitmap case only requires the original source positions for
drawing, create `prepare_for_direct_bitmap_drawing` which is a version
of `prepare_for_direct_mask_drawing` but associates the original source
positions when accepting instead of the mapped mask position. This is
more in keeping with also already existing `prepare_for_path_drawing`
and `prepare_for_drawable_drawing`. This also allows for the removal of
an extra std::vector and its associated allocations.

Bug: 381362590
Change-Id: Ie489c5853be003928652136a85bb68d3ec00ea04
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/932556
Reviewed-by: Brett Somocurcio <brettos@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>