[pdf] Allow simple /StructElem /K
The /StructElem /K (structure element children) is currently always an
array of objects, but it is also allowed to be a single object (without
the array syntax) if there is only a single child. Since this is a
common case at the leaves of the structure element tree, allow emitting
the array of children as a single object if there is only one.

There are a number of other possible uses of this pattern elsewhere in
the specification (though the current generation code does not emit
those constructs at this time). As a result this is done generically
instead of as a one-off when inserting the `kids`.

Change-Id: Ic4ae966b4137525242a9936fd015379acd3822b1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/933716
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>