Allow MSKP benches to throttle frames on GPU.
Currently when timing nanobench on the GPU backends, we introduce an
artificial allowed frame lag to mimic what is happening on a real
device. This way if GPU work is taking a long time, we can't continue to
send infinite frames to be rendered.

Currently this is controlled above the Benchmarks themselves. Each time
after we call draw on the Benchmark, nanobench will submit the work to
the GPU and wait on frames to finish on the GPU if needed.

The issue with MSKPs is that they draw N frames each time the draw()
call is made. So these N frames can be all queued up on the CPU even if
the GPU is still working on previous frames. So we need a way for the
MSKP benchmark to notify when it finishes each frame.

This change adds a callback that the MSKPBench now uses after each
frame. Nanobench will then call endFrame on each Target when this
callback is made.

Change-Id: I5a5e934880dd1423ae15c8855e023b687a4caabe
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/932976
Commit-Queue: Greg Daniel <egdaniel@google.com>
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>