[graphite] Consolidate vulkan ycbcr conversion handling
Leans on ImmutableSamplerInfo to make writing keys simpler and reduce
the API on VulkanYcbcrConversion for making conversions and/or keys.

Removes the ycbcrPackaging namespace, with most of its details now
internal to VulkanYcbcrConversion.

A similar follow-up CL will be applied to the Dawn immutable sampler
tracking.

Change-Id: Id8f2e569d125b3664021bd55f0582387003d3182
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/934376
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>