Make defensive copies of SkPath less frequently when drawing them
SkDevice::drawPath has a curious parameter pathIsMutable. In the
implementation we use const_cast to mutate the passed in variable
(which is a bit...fragile, IMO) instead of making a copy.

I would like to remove that const_cast, perhaps with an API
that takes in SkPath&& so we can use std::move() to pass an
SkPath we created in efficiently when we can and make the
copy to mutate when we cannot. Before doing that, I wanted
to make the behavior more explicit and noticed a few places
where weren't making use of the "save a copy, cast a pointer"
behavior. I didn't notice any major change in performance
when running locally, but maybe we'll see something on Perf
after this lands.

As a follow-up, we can try adding those new function declarations.

Change-Id: I5c2b08812f435ded1830abe65f13ea2ff22bc0cb
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/931831
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>