Combine SkColorShader and SkColor4Shader
SkColorShader now holds an SkColor4f stored in sRGB. The Color
factory that takes a color space now converts to sRGB up front so
that it does not need to be retained.

Fixes color space conversion in Graphite, which was being applied for
SkColor4Shader but not SkColorShader.

Bug: b/310924045
Bug: b/388496775
Change-Id: I5bbebb44699453b14ebf2cd513ee7ecfc8f53cf2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/936438
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>