Use SkCrabbyAvifCodec for HEIC and AVIF on Android
Use SkCrabbyAvifCodec by default for HEIC and AVIF images
on the Android framework.

With this change, we will no longer need/use SkHeifCodec on
Android. I will clean that up in a follow-up CL once this
soaks in for a while without any failures (clean up is tracked
in b/382594683).

Bug: b/364911926
Change-Id: I309f171432d1db65d1310ed29fb7daf808f7abe6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/937516
Reviewed-by: Noelle Scobie <nscobie@google.com>
Commit-Queue: Vignesh Venkat <vigneshv@google.com>