[graphite] Add Scratch shareable concept to ResourceCache
From the ResourceCache's POV, kScratch and kShareable are basically
the same in terms of when it belongs to the fResourceMap. Checks
that had been "shareable == yes" are now updated to "shareable != no".

The main difference is that when requesting kScratch resources, the
caller passes in a set of "unavailable" pointers that is used in the
filter function passed to the fResourceMap.find() call. These
represent scratch resources that have been already checked out by the
caller and it doesn't want them to be aliased with this new request.

Renamed the existing findOrCreateScratchTexture on ResourceProvider
to be findOrCreateNonShareableTexture, and added a new
findOrCreateScratchTexture function that will be used in the follow
up CL by the ScratchResourceManager.

All the different RP texture functions now simply call into the
findOrCreateTexture() call because they don't need distinct key types
anymore (since the shareability is off the key).

Bug: b/387505623
Change-Id: Iefd9dc1cbbcaf1b5995aeacd828bc058edd39e34
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/934836
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>