[graphite] Consolidate path atlas translations into a single place.
We do two translations when rendering a path into an atlas: one to
remove the integer translation from the localToDevice transform,
and one to put it in the correct location in the atlas. This puts those
translations into a single place in both RasterPathAtlas and
ComputePathAtlas to make it clearer.

Bug: b/388809647
Change-Id: I42e778ede328be3e34d00c22f480ca8967052dcd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/938061
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>