[Graphite] Stop letting label back to Scratch in ResourceCache.
When dumping resources or looking at them in a debugging tool it can be
interesting to know whether a Resource is just sitting in the cache or
currently in use. Thus we originally set its label back to Scratch when
it was returned to the cache.

However, after debugging multiple memory related issues, it seems for
valuable to preserve the previous label of how something was used.
Right now we just end up with 75% or more of the Resources just being
called Scratch and it gives no indication of why we may have made it in
the first place.

I think we could consider follow up changes to allow reporting a
Resources status in the cache, but in the short term I think no updating
the label to Scratch is more valuable than doing it.

Change-Id: I40d8c28fcf6dbb2b8a57eb7de3fa316825eab06a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/940442
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>