[graphite] Remove unref mutex and combine ref counts into single atomic
Before we were storing refs in int32_ts, but weren't doing anything
with the sign bit, so for usage and command buffer refs, 31 bits has
proven sufficient. Previously, the ResourceCache would hold between
1 and 2 refs total for a resource. Now that's split into a cache ref
and a return queue ref, each of which can only be set once.

This adds up to 64 bits so everything is stored in a
std::atomic<uint64_t>. With atomic operations the prior value is
returned, which allows state to be tracked based on what that was
before a decrement.

The most complicated change was adding a compare-and-swap/exchange
loop to unref when removing a command or usage ref. This was so that
if it was the last ref preventing reuse, the return queue ref would
be included in the atomic update.

Previously, resources used a bool to say if command buffer refs
would just add usage refs. This is functionally equivalent to
"reusability requires purgeability", which I found slightly simpler
to reason about. The Resource precomputes a bit mask for its ref
count; if this is 0 then it is reusable and can go back to the cache.

Recommended reviewing order:
- The big comment at the top of Resource.h
- The addRef template function
- The removeRef template function
- Resource::unrefReturnQueue()
- Resource::returnToCache()
- ResourceCache
- The other minor updates to subclasses

Bug: b/331589189
Change-Id: Ia30e0845c1bc46322dfb2c03996e6278a58d4a30
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/888809
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>