[SkSL] Add check for unsized array in checkForOutOfRangeLiteral
This fixes a fuzzer bug that created unsized array variable references and then triggered the switch variable-hoisting code. Within the variable-hoisting we assert that arrays have size when getting the slotCount. This CL makes it so that before trying to get slotCount, we simply check if the array is unsized.

This is preferable to not emitting unsized arrays to IR nodes, which can add more errors to the error reporter down the line.

Bug: 388312952
Change-Id: I0977c635a842ae32e368a0ad79ff4a4dac3bc01a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/942956
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>