Try using less stack space for auto blitters
The existing value seemed quite arbitrary. I did a bit of research
into what the actual usage was and based the new value on that.

This is still much bigger than it needs to be in a lot of cases
because users who just use SkDrawBase only seem to have a
BlitterChooser that is SkA8Blitter_Choose, which has a much
simpler (and smaller) use pattern of the stack memory.

I couldn't think of a way to handle that without a bunch of
overrides and duplicated code. Not sure if it's worth that.

Of note, in about 10 seconds of using Chromium, about
20k of these SkAutoBlitterChoose objects were created and
destroyed. Not sure how many of them live concurrently, but
reducing the memory by ~15% might have some meaningful perf
increases.

Change-Id: If2f530beb9fc6eae30f258c8e0575787569addb8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/942996
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>