[graphite] Hook in atlas clip via PaintParams.
Adds NonMSAAClipBlock to encapsulate both analytic clip and atlas clip
data, and a decision point in KeyHelpers to use the analytic clip
alone, or a combination shader.

Also renames any instances of CircularRRectClip to AnalyticClip, to
improve naming consistency and make it less specific -- in case we
want to change it in the future.

Bug: b/388812104
Change-Id: I28170faaecec90ecd349ba1098b44fadcf3c1bac
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/943080
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>