[graphite] Add serialization helpers for the Android-specific Precompile API
This is a somewhat artificial CL in that it doesn't do anything by itself. It mainly functions to split the full Android-specific Precompilation CL into more easily reviewed pieces.

In general I've tried to keep the serialization code separate (in SerializationUtils.*). The exception to this is the TextureInfoData serialization (which required some virtual dispatching) and Caps (which handles virtual dispatch for deserialization).

There are several caveats w/ this CL:
  only Mtl is handled (Vk and Dawn will be forthcoming)
  non-stably-keyed runtime effects are not handled
  versioning is very loosey-goosey right now (forthcoming CLs w/ make it more bullet proof)

Bug: b/390186667
Change-Id: I941eeddb5c217c00ebeb6d46dad2402b58afc7dd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/943936
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>